pyproject.toml
[tool.poetry]
name = "vrp_multi_exec"
version = "0.1.0"
description = ""
authors = ["Maged Motawea <magedmotawea@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
paramiko = "^2.7.2"
Jinja2 = "^2.11.3"
gevent = "^21.1.2"
loguru = "^0.5.3"
click = "^7.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.3"
pdpp = "^0.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
multi_exec = "vrp_multi_exec.entry_points.multi_exec:cli"